Reimplement scope module to use a stack.
