+++ src/analyzer/scope/stack/mod.rs
