+++ src/analyzer/scope/stack/tests/mod.rs
