+++ src/analyzer/scope/tests/statement/mod.rs
@@ -3,8 +3,6 @@ use crate::analyzer::SemanticAnalyzer;
 use crate::ast::expression::Expr;
 use crate::ast::statement::Stmt;
 
-mod expression;
-
 #[test]
 fn expression_stmt_should_return_ok() {
     let stmts = vec![Stmt::Expression(Expr::Primary(obj_bool!(true)))];
@@ -161,6 +159,67 @@ fn while_statement_should_return_self() {
             Expr::Primary(obj_bool!(true)),
         )),
     )];
+
+    assert_eq!(Ok(output), ScopeAnalyzer::new().analyze(input))
+}
+
+#[test]
+fn scope_should_shade_correctly() {
+    let input = vec![
+        Stmt::Declaration(
+            identifier_name!("a"),
+            Expr::Primary(obj_str!("global".to_string())),
+        ),
+        Stmt::Block(vec![
+            Stmt::Function(
+                identifier_name!("showA"),
+                vec![],
+                Box::new(Stmt::Block(vec![Stmt::Print(Expr::Variable(
+                    identifier_name!("a"),
+                ))])),
+            ),
+            Stmt::Expression(Expr::Call(
+                Box::new(Expr::Variable(identifier_name!("showA"))),
+                vec![],
+            )),
+            Stmt::Declaration(
+                identifier_name!("a"),
+                Expr::Primary(obj_str!("block".to_string())),
+            ),
+            Stmt::Expression(Expr::Call(
+                Box::new(Expr::Variable(identifier_name!("showA"))),
+                vec![],
+            )),
+        ]),
+    ];
+
+    let output = vec![
+        Stmt::Declaration(
+            identifier_id!(0),
+            Expr::Primary(obj_str!("global".to_string())),
+        ),
+        Stmt::Block(vec![
+            Stmt::Function(
+                identifier_id!(1),
+                vec![],
+                Box::new(Stmt::Block(vec![Stmt::Print(Expr::Variable(
+                    identifier_id!(0),
+                ))])),
+            ),
+            Stmt::Expression(Expr::Call(
+                Box::new(Expr::Variable(identifier_id!(1))),
+                vec![],
+            )),
+            Stmt::Declaration(
+                identifier_id!(2),
+                Expr::Primary(obj_str!("block".to_string())),
+            ),
+            Stmt::Expression(Expr::Call(
+                Box::new(Expr::Variable(identifier_id!(1))),
+                vec![],
+            )),
+        ]),
+    ];
 
     assert_eq!(Ok(output), ScopeAnalyzer::new().analyze(input))
 }
